requirements.txt
dnspython==2.3.0
pika==1.3.1
pika-stubs==0.1.3
pyfiglet==1.0.2
pymongo==4.3.3
PyYAML==6.0
termcolor==2.5.0
uuid==1.30
